+++ sakura_core/doc/CLineComment.cpp
@@ -47,7 +47,7 @@ void CLineComment::CopyTo( const int n, const wchar_t* buffer, int nCommentPos )
 	}
 }
 
-bool CLineComment::Match( int nPos, const CStringRef& cStr ) const
+bool CLineComment::Match( ssize_t nPos, const CStringRef& cStr ) const
 {
 	int i;
 	for ( i=0; i<COMMENT_DELIMITER_NUM; i++ ){
