+++ sakura_core/view/colors/CColor_KeywordSet.h
@@ -16,11 +16,11 @@ public:
 	CColor_KeywordSet();
 	EColorIndexType GetStrategyColor() const override{ return (EColorIndexType)(COLORIDX_KEYWORD1 + m_nKeywordIndex); }
 	void InitStrategyStatus() override{ m_nCOMMENTEND = 0; }
-	bool BeginColor(const CStringRef& cStr, int nPos) override;
-	bool EndColor(const CStringRef& cStr, int nPos) override;
+	bool BeginColor(const CStringRef& cStr, ssize_t nPos) override;
+	bool EndColor(const CStringRef& cStr, ssize_t nPos) override;
 	bool Disp() const override{ return true; }
 private:
 	int m_nKeywordIndex;
-	int m_nCOMMENTEND;
+	ssize_t m_nCOMMENTEND;
 };
 #endif /* SAKURA_CCOLOR_KEYWORDSET_AAB04E86_EB95_447C_B07B_AB44395B2F7F_H_ */
